split tests up
